+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Builtin function `read_entire_file(filename: String)`
 - String indexing
+- String comparison
 
 ### Changed
 ### Fixed